backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Authentication and Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7
bcrypt==4.1.2

# Database
django-mssql-backend==2.8.1
pyodbc==5.0.1

# File Processing and OCR
Pillow==10.1.0
python-magic==0.4.27
PyPDF2==3.0.1
pdf2image==1.16.3
anthropic==0.7.8

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# Validation and Serialization
marshmallow==3.20.1
cerberus==1.3.5

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3
celery==5.3.4
redis==5.0.1

# Logging and Monitoring
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0